+++ pool/pgpoolstorage/pgpoolstorage.go
@@ -802,33 +802,6 @@ func (p *PostgresPoolStorage) GetAllAddressesBlocked(ctx context.Context) ([]com
 	return addrs, nil
 }
 
-// GetAllAddressesWhitelisted get all addresses whitelisted
-func (p *PostgresPoolStorage) GetAllAddressesWhitelisted(ctx context.Context) ([]common.Address, error) {
-	sql := `SELECT addr FROM pool.whitelisted`
-
-	rows, err := p.db.Query(ctx, sql)
-	if err != nil {
-		if errors.Is(err, pgx.ErrNoRows) {
-			return nil, nil
-		} else {
-			return nil, err
-		}
-	}
-	defer rows.Close()
-
-	var addrs []common.Address
-	for rows.Next() {
-		var addr string
-		err := rows.Scan(&addr)
-		if err != nil {
-			return nil, err
-		}
-		addrs = append(addrs, common.HexToAddress(addr))
-	}
-
-	return addrs, nil
-}
-
 // GetEarliestProcessedTx gets the earliest processed tx from the pool. Mainly used for cleanup
 func (p *PostgresPoolStorage) GetEarliestProcessedTx(ctx context.Context) (common.Hash, error) {
 	const getEarliestProcessedTxnFromTxnPool = `SELECT hash
